+++ sentry-spring/src/test/kotlin/io/sentry/spring/tracing/SentryTracingFilterTest.kt
@@ -15,7 +15,7 @@ import io.sentry.SentryOptions
 import io.sentry.SentryTracer
 import io.sentry.SpanId
 import io.sentry.SpanStatus
-import io.sentry.TraceState
+import io.sentry.TraceContext
 import io.sentry.TransactionContext
 import io.sentry.protocol.SentryId
 import org.assertj.core.api.Assertions.assertThat
@@ -84,7 +84,7 @@ class SentryTracingFilterTest {
                 assertThat(it.contexts.trace!!.status).isEqualTo(SpanStatus.OK)
                 assertThat(it.contexts.trace!!.operation).isEqualTo("http.server")
             },
-            anyOrNull<TraceState>(),
+            anyOrNull<TraceContext>(),
             anyOrNull(),
             anyOrNull()
         )
@@ -100,7 +100,7 @@ class SentryTracingFilterTest {
             check {
                 assertThat(it.contexts.trace!!.status).isEqualTo(SpanStatus.INTERNAL_ERROR)
             },
-            anyOrNull<TraceState>(),
+            anyOrNull<TraceContext>(),
             anyOrNull(),
             anyOrNull()
         )
@@ -116,7 +116,7 @@ class SentryTracingFilterTest {
             check {
                 assertThat(it.contexts.trace!!.status).isNull()
             },
-            anyOrNull<TraceState>(),
+            anyOrNull<TraceContext>(),
             anyOrNull(),
             anyOrNull()
         )
@@ -132,7 +132,7 @@ class SentryTracingFilterTest {
             check {
                 assertThat(it.contexts.trace!!.parentSpanId).isNull()
             },
-            anyOrNull<TraceState>(),
+            anyOrNull<TraceContext>(),
             anyOrNull(),
             anyOrNull()
         )
@@ -149,7 +149,7 @@ class SentryTracingFilterTest {
             check {
                 assertThat(it.contexts.trace!!.parentSpanId).isEqualTo(parentSpanId)
             },
-            anyOrNull<TraceState>(),
+            anyOrNull<TraceContext>(),
             anyOrNull(),
             anyOrNull()
         )
@@ -182,7 +182,7 @@ class SentryTracingFilterTest {
             check {
                 assertThat(it.status).isEqualTo(SpanStatus.INTERNAL_ERROR)
             },
-            anyOrNull<TraceState>(),
+            anyOrNull<TraceContext>(),
             anyOrNull(),
             anyOrNull()
         )